handling error

+++ src/queries/api/remboursement_queries.ts
@@ -7,7 +7,7 @@ remboursementRouter.use(bodyParser.json());
 remboursementRouter.get('/', (request, response) => {
     pool.query('select * from remboursements;', (error, results) => {
         if (error) {
-            throw error
+            response.status(400).json({message: error})
         }
         response.status(200).json(results.rows)
     })
@@ -16,7 +16,7 @@ remboursementRouter.get('/', (request, response) => {
 remboursementRouter.get('/ofgroup/:id', (request, response) => {
     pool.query('select utilisateur,part,idgroupe from remboursements where idgroupe = $1;', [request.params.id], (error, results) => {
         if (error) {
-            throw error
+            response.status(400).json({message: error})
         }
         response.status(200).json(results.rows)
     })
@@ -26,7 +26,7 @@ remboursementRouter.get('/ofgroup/ofuser/:id', (request, response) => {
     const { email } = request.body;
     pool.query('select part,iddepense from remboursements where idgroupe = $1 and utilisateur=$2;', [request.params.id, email], (error, results) => {
         if (error) {
-            throw error
+            response.status(400).json({message: error})
         }
         response.status(200).json(results.rows)
     })
@@ -36,9 +36,13 @@ remboursementRouter.get('/ofgroup/ofuser/:id', (request, response) => {
 
 remboursementRouter.post('/', (request, response) => {
     const { iddepense, idgroupe, utilisateur, part } = request.body
+    if(!iddepense || !part|| !utilisateur || !idgroupe){
+        response.status(400).json({message: "Bad request: can't create "})
+        return
+    }
     pool.query('insert into remboursements (iddepense, idgroupe, utilisateur, part) values ($1,$2,$3,$4);', [iddepense, idgroupe, utilisateur, part], (error, results) => {
         if (error) {
-            console.log(error); 
+            response.status(400).json({message: error}) 
         }
         response.status(200).json(results.rows)
     })
@@ -46,9 +50,13 @@ remboursementRouter.post('/', (request, response) => {
 
 remboursementRouter.delete('/:id', (request, response) => {
     const { iddepense } = request.body
+    if(!iddepense){
+        response.status(400).json({message: "Bad request:can't delete depense"})
+        return
+    }
     pool.query('delete from remboursements where iddepense=$1 and idgroupe=$2;', [iddepense, request.params.id], (error, results) => {
         if (error) {
-            throw error
+            response.status(400).json({message: error})
         }
         response.status(200).json(results.rows)
     })
@@ -59,7 +67,7 @@ remboursementRouter.put('/:id', (request, response) => {
     const { part, iddepense } = request.body
     pool.query('update remboursements set part = $1 where iddepense=$2 and idgroupe=$3;', [part, iddepense, request.params.id], (error, results) => {
         if (error) {
-            throw error
+            response.status(400).json({message: error})
         }
         response.status(200).json(results.rows)
     })
@@ -68,7 +76,7 @@ remboursementRouter.put('/:id', (request, response) => {
 remboursementRouter.get('/detail/:id',(request,response)=>{
     pool.query("select sum(d.montant*r.part) as total , d.utilisateur, r.utilisateur as emprunteur from depenses d join remboursements r on r.iddepense = d.id where d.idgroupe = $1 group by d.utilisateur ,r.utilisateur", [request.params.id],(error,results)=>{
         if (error) {
-            throw error;
+            response.status(400).json({message: error})
         }
         response.status(200).json(results.rows);
     })
